burger controller all and index view complete

+++ db/burgers.sql
@@ -14,23 +14,10 @@ name VARCHAR(255)
 CREATE TABLE burgers(
 id SERIAL8 primary key,
 name VARCHAR(255),
-price FLOAT,
+price DECIMAL(3,2),
 restaurant VARCHAR(255),
 deal VARCHAR (255),
 day_id INT8 references days(id)
 );
 
--- CREATE TABLE deals(
--- id SERIAL8 primary key,
--- name VARCHAR(255),
--- description TEXT,
--- burger_id INT8 references burgers(id),
--- day_id INT8 references days(id)
--- );
 
--- CREATE TABLE restaurants(
--- id SERIAL8 primary key,
--- name VARCHAR(255),
--- burger_id INT8 references burgers(id),
--- deal_id INT8 references deals(id)
--- );
